Source files, data for r=1.142
Teleportation and amplification scheme following https://journals.aps.org/prl/abstract/10.1103/PhysRevLett.128.160501. State to be amplified is single mode squeezed vacuum with r=1.142. This is source code and data for gain parameters ranging from 1/4 to 128, with 10^6 shots each. For gains of 1/2, 1, 2, 4, and 8 the experiment was run a second time as _v2 to collect more shots.